Remove krb5-identity extensions; add coverage report
It is not feasible to support the krb5 idenity extensions
because the C libraries write the values to thread local storage
which is not stable on Golang due to goroutines moving between
OS threads.

Don't rely on the init() function ordering for inspecting optional
symbols - we do this when first used instead.

Fix FreeBSD tests