Fixed latent pinning bug related to OID sets
The Go code that built up OID sets was faults because the resultant
array of OID pointers was not from a single allocation and so
turning that back into a Go slice would fail.

We moved to using the C GSSAPI Oid-set builder instead.

Additionally use the OS/arch in the toolbin path to avoid
clashes when testing on multiple platforms with a single checkout.