Don't set credential interface value to concrete nil
Previously we were returning gssapi.SecContextInfoPartial.DelegatedCredential
as a nil Credential.  This means that the return (interface) value is no
longer nil, but carries a nil pointer to the concrete type.

Instead only set the interface value at all if the concrete type is not
nil.  Then we can check SecContextInfoPartial.DelegatedCredential == nil
reliably in caller code.